Move call_string to utils
