+++ plasmapy/classes/sources/plasmablob.py
@@ -10,8 +10,7 @@ from plasmapy.formulary.dimensionless import quantum_theta
 from plasmapy.formulary.collisions import coupling_parameter
 from plasmapy.atomic import particle_mass
 
-from plasmapy.utils import CouplingWarning
-from plasmapy.utils.pytest_helpers.error_messages import call_string
+from plasmapy.utils import CouplingWarning, call_string
 
 from plasmapy.classes import GenericPlasma
