+++ plasmapy/utils/__init__.py
@@ -16,3 +16,4 @@ from plasmapy.utils.exceptions import (
     RelativityError,
     RelativityWarning,
 )
+from .error_messages import call_string
