+++ plasmapy/utils/pytest_helpers/__init__.py
@@ -4,8 +4,6 @@ from plasmapy.utils.pytest_helpers.pytest_helpers import (
     assert_can_handle_nparray,
 )
 
-from .error_messages import call_string
-
 from .exceptions import (
     InconsistentTypeError,
     UnexpectedResultError,
